rythubadi-AI-utils/requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
sentence-transformers
chromadb
faiss-cpu
wikipedia
arxiv
langchain_openai
langchainhub